+++ apps/user/src/components/form/ProfileUpload/ProfileUpload.tsx
@@ -1,16 +1,8 @@
 import { color, font } from '@maru/theme';
 import { Button, Column } from '@maru/ui';
-import {
-    ChangeEventHandler,
-    Dispatch,
-    SetStateAction,
-    useEffect,
-    useState,
-    useRef,
-    DragEvent,
-} from 'react';
+import { ChangeEventHandler, Dispatch, SetStateAction, useState, useRef, DragEvent } from 'react';
 import { UserInfo } from '@/types/form';
-import { useUploadProfileImageMutation } from '@/services/form/지원자정보/mutations';
+import { useUploadProfileImageMutation } from '@/services/form/mutations';
 import styled from 'styled-components';
 
 interface PropsType {
