+++ apps/user/src/services/form/mutations.ts
@@ -1,6 +1,7 @@
-import { Form } from '@/types/form';
+import { Form, UserInfo } from '@/types/form';
+import { Dispatch, SetStateAction } from 'react';
 import { useMutation } from 'react-query';
-import { submitDraftForm } from './api';
+import { submitDraftForm, uploadProfileImage } from './api';
 
 export const useSubmitDraftFormMutation = (formData: Form) => {
     const { mutate: submitDraftFormMutate, ...restMutation } = useMutation({
@@ -16,4 +17,18 @@ export const useSubmitDraftFormMutation = (formData: Form) => {
     });
 
     return { submitDraftFormMutate, restMutation };
+};
+
+export const useUploadProfileImageMutation = (setUserInfo: Dispatch<SetStateAction<UserInfo>>) => {
+    const { mutate: uploadProfileImageMutate, ...restMutation } = useMutation({
+        mutationFn: (image: FormData) => uploadProfileImage(image),
+        onSuccess: (res) => {
+            setUserInfo((prev) => ({ ...prev, identificationPictureUri: res.data.data.url }));
+        },
+        onError: (err) => {
+            console.log(err);
+        },
+    });
+
+    return { uploadProfileImageMutate, ...restMutation };
 };
