+++ apps/user/src/services/form/queries.ts
@@ -1,17 +1,11 @@
 import KEY from '@/constants/key';
 import { useQuery } from 'react-query';
-import { getSchoolList } from './api';
-
-interface SchoolList {
-    code: string;
-    name: string;
-    location: string;
-}
+import { schoolList } from './api';
 
 export const useSchoolListQuery = (school: string) => {
     return useQuery<SchoolList[]>({
         queryKey: [KEY.FORM_SCHOOL_LIST, school],
-        queryFn: () => getSchoolList(school),
+        queryFn: () => schoolList(school),
         initialData: [],
     });
 };
