--- apps/user/src/services/form/출신학교및학력/api.ts
@@ -1,7 +1,0 @@
-import { maru } from '@/apis/instance/instance';
-
-export const getSchoolList = async (school: string) => {
-    const { data } = await maru.get(`/school?q=${school}`);
-
-    return data.dataList;
-};
